Move deserialization into PeerMessage to distinct variants correctly
