+++ aquadoggo/src/replication/announcement.rs
@@ -2,14 +2,12 @@
 
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use p2panda_rs::Validate;
-use serde::de::Visitor;
 use serde::ser::SerializeSeq;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 
-use crate::replication::{MessageType, SchemaIdSet, ANNOUNCE_TYPE, REPLICATION_PROTOCOL_VERSION};
+use crate::replication::{SchemaIdSet, ANNOUNCE_TYPE, REPLICATION_PROTOCOL_VERSION};
 
-/// u64 timestamp from UNIX epoch until now.
+/// U64 timestamp from UNIX epoch until now.
 pub fn now() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
@@ -40,7 +38,7 @@ pub type ProtocolVersion = u64;
 
 /// Message which can be used to send announcements over the wire.
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct AnnouncementMessage(ProtocolVersion, Announcement);
+pub struct AnnouncementMessage(pub ProtocolVersion, pub Announcement);
 
 impl AnnouncementMessage {
     pub fn new(announcement: Announcement) -> Self {
@@ -70,76 +68,10 @@ impl Serialize for AnnouncementMessage {
     }
 }
 
-impl<'de> Deserialize<'de> for AnnouncementMessage {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct MessageVisitor;
-
-        impl<'de> Visitor<'de> for MessageVisitor {
-            type Value = AnnouncementMessage;
-
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("p2panda announce message")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-            where
-                A: serde::de::SeqAccess<'de>,
-            {
-                let message_type: MessageType = seq.next_element()?.ok_or_else(|| {
-                    serde::de::Error::custom("missing message type in announce message")
-                })?;
-
-                if message_type != ANNOUNCE_TYPE {
-                    return Err(serde::de::Error::custom(
-                        "invalid message type for announce message",
-                    ));
-                }
-
-                let protocol_version: ProtocolVersion = seq.next_element()?.ok_or_else(|| {
-                    serde::de::Error::custom("missing protocol version in announce message")
-                })?;
-
-                let timestamp: u64 = seq.next_element()?.ok_or_else(|| {
-                    serde::de::Error::custom("missing timestamp in announce message")
-                })?;
-
-                let supported_schema_ids: SchemaIdSet = seq.next_element()?.ok_or_else(|| {
-                    serde::de::Error::custom("missing target set in announce message")
-                })?;
-                supported_schema_ids.validate().map_err(|_| {
-                    serde::de::Error::custom("invalid target set in announce message")
-                })?;
-
-                if let Some(items_left) = seq.size_hint() {
-                    if items_left > 0 {
-                        return Err(serde::de::Error::custom(
-                            "too many fields for announce message",
-                        ));
-                    }
-                };
-
-                Ok(AnnouncementMessage(
-                    protocol_version,
-                    Announcement {
-                        supported_schema_ids,
-                        timestamp,
-                    },
-                ))
-            }
-        }
-
-        deserializer.deserialize_seq(MessageVisitor)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use ciborium::cbor;
-    use ciborium::value::{Error, Value};
-    use p2panda_rs::serde::{deserialize_into, serialize_from, serialize_value};
+    use p2panda_rs::serde::{serialize_from, serialize_value};
     use rstest::rstest;
 
     use crate::replication::SchemaIdSet;
@@ -154,39 +86,5 @@ mod tests {
             serialize_from(AnnouncementMessage::new(announcement.clone())),
             serialize_value(cbor!([0, 1, announcement.timestamp, supported_schema_ids]))
         );
-    }
-
-    #[rstest]
-    fn deserialize(#[from(random_schema_id_set)] supported_schema_ids: SchemaIdSet) {
-        assert_eq!(
-            deserialize_into::<AnnouncementMessage>(&serialize_value(cbor!([
-                0,
-                1,
-                12345678,
-                supported_schema_ids
-            ])))
-            .unwrap(),
-            AnnouncementMessage::new(Announcement {
-                timestamp: 12345678,
-                supported_schema_ids,
-            })
-        );
-    }
-
-    #[rstest]
-    #[should_panic(expected = "missing message type in announce message")]
-    #[case::missing_version(cbor!([]))]
-    #[should_panic(expected = "missing protocol version in announce message")]
-    #[case::missing_version(cbor!([0]))]
-    #[should_panic(expected = "missing timestamp in announce message")]
-    #[case::missing_timestamp(cbor!([0, 122]))]
-    #[should_panic(expected = "too many fields for announce message")]
-    #[case::too_many_fields(cbor!([0, 1, 0, ["schema_field_definition_v1"], "too much"]))]
-    fn deserialize_invalid_messages(#[case] cbor: Result<Value, Error>) {
-        // Check the cbor is valid
-        assert!(cbor.is_ok());
-
-        // We unwrap here to cause a panic and then test for expected error stings
-        deserialize_into::<AnnouncementMessage>(&serialize_value(cbor)).unwrap();
     }
 }
